A Class which represents a guild object

A Class which represents a guild object

Guild(data)

Slots

id

guild id

name

guild name (2-100 characters)

icon

icon hash

splash

splash hash

owner

whether or not the user is the owner of the guild

owner_id

id of owner

permissions

total permissions for the user in the guild (does not include channel overrides)

region

voice region id for the guild

afk_channel_id

id of afk channel

afk_timeout

afk timeout in seconds

embed_enabled

is this guild embeddable (e.g. widget)

embed_channel_id

if not null, the channel id that the widget will generate an invite to

verification_level

verification level required for the guild

default_message_notifications

default message notifications level

explicit_content_filter

explicit content filter level

roles

roles in the guild

emojis

custom guild emojis

features

enabled guild features

mfa_level

required MFA level for the guild

application_id

application id of the guild creator if it is bot-created

widget_enabled

whether or not the server widget is enabled

widget_channel_id

the channel id for the server widget

system_channel_id

the id of the channel to which system messages are sent

joined_at

when this guild was joined at

large

whether this is considered a large guild

unavailable

is this guild unavailable

member_count

total number of members in this guild

voice_stated

(without the guild_id key)

members

users in the guild

channels

channels in the guild

presences

presences of the users in the guild

max_presences

the maximum amount of presences for the guild (the default value, currently 5000, is in effect when null is returned)

max_members

the maximum amount of members for the guild

vanity_url_code

the vanity url code for the guild

description

the description for the guild

banner

banner hash

premium_tier

premium_tier (SEE: https://discordapp.com/developers/docs/resources/guild#guild-object-premium-tier)

premium_subscription_count

the total number of users currently boosting this server